refactor: 값 전달 객체로 사용할 수 있는 객체를 domain/dto 패키지로 이동
