--- src/main/kotlin/com/example/hhplus_ecommerce/api/order/response/OrderItem.kt
@@ -1,8 +1,0 @@
-package com.example.hhplus_ecommerce.api.order.response
-
-class OrderItem(
-	val productId: Long,
-	val quantity: Int,
-	val price: Int
-) {
-}
