+++ src/main/kotlin/com/example/hhplus_ecommerce/api/order/response/OrderResponse.kt
@@ -1,5 +1,6 @@
 package com.example.hhplus_ecommerce.api.order.response
 
+import com.example.hhplus_ecommerce.domain.order.dto.OrderItemDetail
 import java.time.LocalDate
 
 class OrderResponse(
@@ -7,6 +8,6 @@ class OrderResponse(
 	val userId: Long,
 	val orderDate: LocalDate,
 	val totalPrice: Int,
-	val orderItems: List<OrderItem>
+	val orderItems: List<OrderItemDetail>
 ) {
 }
