+++ src/main/kotlin/com/example/hhplus_ecommerce/domain/order/dto/OrderItemDetail.kt
@@ -1,0 +1,8 @@
+package com.example.hhplus_ecommerce.domain.order.dto
+
+class OrderItemDetail(
+	val productId: Long,
+	val quantity: Int,
+	val price: Int
+) {
+}
